Language.SQL.Keyword.Concat

Description

Concatinations on Keyword types

List concatination functions

Functions to concatinate Keyword list.

unwords' :: [Keyword] -> Keyword

Concatinate Keyword list like unwords on String list.

sepBy :: [Keyword] -> Keyword -> Keyword

Separate Keyword list with delimiter Keyword and concatinate into one Keyword.

parenSepBy :: [Keyword] -> Keyword -> Keyword

Do sepBy and enclose by paren

Binary operators

Binary operators on SQL. Result is concatinated into one Keyword.

defineBinOp :: Keyword -> Keyword -> Keyword -> Keyword

Define binary operator on Keyword type. Result is delimited by whitespace like unwords on String list.

strBinOp :: (Keyword -> Keyword -> Keyword) -> String -> String -> String

Define binary operator of string from Keyword binary operator.

as :: Keyword -> Keyword -> Keyword

Binary operator for SQL name alias.

(<.>) :: Keyword -> Keyword -> Keyword

Binary operator to create qualified name on SQL.

(|*|) :: Keyword -> Keyword -> Keyword infixr 1

Binary operator to create comma separated words.

(.||.) :: Keyword -> Keyword -> Keyword infixr 5

Binary operator for SQL string expression concatination.

(.=.) :: Keyword -> Keyword -> Keyword infixr 4

Binary eq operator for SQL expression.

(.<.) :: Keyword -> Keyword -> Keyword infixr 4

Binary lt operator for SQL expression.

(.<=.) :: Keyword -> Keyword -> Keyword infixr 4

Binary le operator for SQL expression.

(.>.) :: Keyword -> Keyword -> Keyword infixr 4

Binary gt operator for SQL expression.

(.>=.) :: Keyword -> Keyword -> Keyword infixr 4

Binary ge operator for SQL expression.

(.<>.) :: Keyword -> Keyword -> Keyword infixr 4

Binary not eq operator for SQL expression.

and :: Keyword -> Keyword -> Keyword infixr 3

Binary AND operator for SQL boolean expression.

or :: Keyword -> Keyword -> Keyword infixr 2

Binary OR operator for SQL boolean expression.

in' :: Keyword -> Keyword -> Keyword infix 4

Binary IN operator for SQL.

(<++>) :: Keyword -> Keyword -> Keyword

Directly concatinate SQL string without whitespaces.

fold

Arguments

:: (Keyword -> Keyword -> Keyword)
Binary operator used in fold

-> [Keyword]
List to fold

-> Keyword
Result

Fold operation using binary operator with empty result of zero length case.

Unary operator

defineUniOp :: Keyword -> Keyword -> Keyword

Define unary operator on Keyword type represeted by specified Keyword. Result is delimited by whitespace like unwords on String list.

paren :: Keyword -> Keyword

Uni operator to create Parend words.

strUniOp :: (Keyword -> Keyword) -> String -> String

Define uni operator of string from Keyword uni operator.